src/node/requirements.txt
grpcio==1.78.0
grpcio-tools==1.78.0
protobuf==6.33.5
setuptools==82.0.1
typing_extensions==4.15.0
